backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.104.1
h11==0.14.0
idna==3.6
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
pypng==0.20220715.0
qrcode==7.4.2
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
